+++ extensions/fullscreentext/CHANGELOG.md
@@ -1,3 +1,10 @@
-# Full Screen Text Changelog
+# Fullscreen Text Changelog
+
+## [New commands] - {PR_MERGE_DATE}
+
+* Add Fullscreen Clipboard command
+* Add Fullscreen Selection command
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+* Add Fullscreen Text command